requirements.txt
fastapi
uvicorn
httpx 
requests
openai
python-dotenv
pydantic
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0.post1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.104
marshmallow==3.22.0
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.42.0
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
tenacity==8.5.0
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
